feat : login_dividing components_done
