+++ frontend/src/components/Join/JoinWrapper.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import styled from 'styled-components';
+
+export default function JoinWrapper() {
+  const JoinWrapper = styled.div`
+    width: 100%;
+    height: 542px;
+    background: #ffffff;
+    border: 1px solid #dbdbdb;
+    margin: 30px 0 10px 0;
+    padding: 0 40px;
+    box-sizing: border-box;
+  `;
+  return <JoinWrapper />;
+}
